feat: 매니저 권한 인터셉터를 추가하고, "/api/manager/**" 에서 동작하도록 추가
- fix: 매니저 권한 인터셉터 추가로 인한 기존 Authority가 Manager 인지 검증하는 로직을 삭제

- refactor: 모든 Service의 메서드 리팩토링, Service 객체에서 타 Service 객체를 의존성 주입하는 것이 아닌 Mapper 주입으로 변경

- fix: 기존 모든 Post 관련 도메인 객체의 정적 변수 삭제

- feat: Inquiry 도메인 객체에서 암호화된 패스워드를 필드에 할당 후 반환하는 메서드 추가

- fix: Manager 도메인 객체에서 고유 식별자를 의미하는 필드의 데이터 타입을 Long에서 Integer로 변경

+++ src/main/java/com/now/core/admin/manager/domain/Manager.java
@@ -19,7 +19,7 @@ public class Manager {
     /**
      * 매니저의 고유 식별자
      */
-    private final Long managerIdx;
+    private final Integer managerIdx;
 
     /**
      * 매니저의 아이디
